+++ src/Config.php
@@ -538,40 +538,12 @@ class Config extends CommonDBTM
             return;
         }
 
-        echo "<form name='form' action=\"" . Toolbox::getItemTypeFormURL(__CLASS__) . "\" method='post' data-track-changes='true'>";
-        echo "<div class='card' id='tabsbody'>";
-        echo "<table class='tab_cadre_fixe'>";
-        echo "<tr><th colspan='4'>" . __('Authentication') . "</th></tr>";
-
-        echo "<tr class='tab_bg_2'>";
-        echo "<td width='30%'>" . __('Automatically add users from an external authentication source') .
-           "</td><td width='20%'>";
-        Dropdown::showYesNo("is_users_auto_add", $CFG_GLPI["is_users_auto_add"]);
-        echo "</td><td width='30%'>" . __('Add a user without accreditation from a LDAP directory') .
-           "</td><td width='20%'>";
-        Dropdown::showYesNo("use_noright_users_add", $CFG_GLPI["use_noright_users_add"]);
-        echo "</td></tr>";
-
-        echo "<tr class='tab_bg_2'>";
-        echo "<td> " . __('Action when a user is deleted from the LDAP directory') . "</td><td>";
-        AuthLDAP::dropdownUserDeletedActions($CFG_GLPI["user_deleted_ldap"]);
-        echo "</td><td> " . __('Action when a user is restored in the LDAP directory') . "</td><td>";
-        AuthLDAP::dropdownUserRestoredActions($CFG_GLPI["user_restored_ldap"]);
-        echo "</td></tr>";
-
-        echo "<tr class='tab_bg_2'>";
-        echo "<td> " . __('GLPI server time zone') . "</td><td>";
-        Dropdown::showGMT("time_offset", $CFG_GLPI["time_offset"]);
-        echo "</td><td></td></tr>";
-
-        echo "<tr class='tab_bg_2'>";
-        echo "<td colspan='4' class='center'>";
-        echo "<input type='submit' name='update_auth' class='btn btn-primary' value=\"" . _sx('button', 'Save') .
-           "\">";
-        echo "</td></tr>";
-
-        echo "</table></div>";
-        Html::closeForm();
+        $twig = TemplateRenderer::getInstance();
+        $twig->display('pages/setup/authentication/setup.html.twig', [
+            'token'                   => Session::getNewCSRFToken(),
+            'action'                  => Toolbox::getItemTypeFormURL(__CLASS__),
+            'CFG_GLPI'                => $CFG_GLPI,
+        ]);
     }